Backend/API/requirements.txt
flask-sqlalchemy
PyJWT
flask_cors
numpy
opencv-python
flask
scikit-learn
insightface
onnxruntime
reportlab